Change ownership of threads (#303)
* make all OpenMP IR names consistent

* thread ref to ptr

* ThreadTrace own child threads

Instead of ProgramTrace taking ownership of all threads, each thread
takes ownership of the threads it spawns. This way threads don't need to
add themselves to the ProgramTrace list of threads. Each thread records
its own children, and at the end program trace can traverse the tree of
threads and construct a list of all threads.